+++ src/popover-example/popover.css
@@ -22,6 +22,15 @@
   background: transparent;
 }
 
+.popover_arrow {
+  width: 0;
+  height: 0;
+  border-left: 5px solid transparent;
+  border-right: 5px solid transparent;
+
+  border-bottom: 5px solid black;
+}
+
 /* 
 .popover_overlay:before {
   content: "";
